requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
sqlparse==0.4.3
tzdata==2022.4
requests
numpy_financial
python-dateutil